dj_ms/requirements.txt

celery==5.3.5
Django==4.1.13
django-extensions==3.2.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
flower==2.0.1
pika==1.3.2
pymongo==3.12.3
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0

